Change sea_route to map overlay without face
Change sea routes from ordinary ground tiles to an invisible overlay.
Because they have no face, the new sea routes are never sent to the
client and are only visible in the map editor.

Add several varieties of the sea_route tile, in case multiple sea routes
overlap on the same map square.